chore(release): Sam-I-Am Release

+++ test/models.py
@@ -113,7 +113,7 @@ class Test(Node):
     def new_key(self, value):
         self._set_property('new_key', value)
 
-    @pg_property(long, str)
+    @pg_property(int, str)
     def timestamp(self, value):
         self._set_property('timestamp', value)
